calendar_modifier/requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
ics